Agregar metodos para calcular presupuestos
Agregar metodos para calcular presupuestos

+++ Windows/src/Presupuestos.java
@@ -1,20 +1,20 @@
 
 public class Presupuestos {
 	
-	public int totalPresupuesto(int uno, int dos, int tres, int cuatro, int cinco, int seis, int siete, int ocho, int nueve, int diez) {
-		int total;
+	public double totalPresupuesto(double uno, double dos, double tres, double cuatro, double cinco, double seis, double siete, double ocho, double nueve, double diez) {
+		double total;
 		
 		total = uno + dos + tres + cuatro + cinco + seis + siete + ocho + nueve + diez;
 		
 		return total;
 	}
 	
-	public int porcentajePresupuesto(int totalPresupuesto, int totalIngresos) {
+	public double porcentajePresupuesto(double totalPresupuesto, double totalIngresos) {
 		double porcentaje;
 		totalIngresos += 1;
 		
 		porcentaje = (totalPresupuesto * 100) / totalIngresos; 
 
-		return (int) Math.round(porcentaje) ;
+		return Math.round(porcentaje) ;
 	}
 }
